Support sites without portal_quickinstaller.
We can use get_installer in Plone 5.1 migrations.
In earlier version we will keep using the portal_quickinstaller,
because get_installer is not available.

In a getToolByName call for portal_quickinstaller, we will pass a
default of None, so an upgrade can still work if the portal tool has
already been removed.

In shared utility and base code, we will try to import get_installer,
and fall back on the previous implementation.

https://github.com/plone/Products.CMFPlone/issues/1340